% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/column_stats.R
\name{column_stats}
\alias{column_stats}
\title{\subsection{Obtain summary statistics of column(s) including count, mean, median, mode, Q1, Q3,
variance, standard deviation, correlation, and covariance in table format.
Parameters}{
}}
\usage{
column_stats()
}
\arguments{
\item{data}{A data set from which columns will be selected}

\item{columns}{A list of columns for which to obtain summary stats, correlation matrix, and covariance matrix
(if > 1 column arguments used)}
}
\value{
three pd.DataFrames
}
\description{
\subsection{Obtain summary statistics of column(s) including count, mean, median, mode, Q1, Q3,
variance, standard deviation, correlation, and covariance in table format.
Parameters}{
}
}
