% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_histogram.R
\name{plot_histogram}
\alias{plot_histogram}
\title{Creates histograms for numerical features within a dataframe using ggplot2}
\usage{
plot_histogram(data, columns, num_bins = 30)
}
\arguments{
\item{data}{A dataframe}

\item{columns}{A list of numerical features for which to create histograms,
or by default will plot all numerical features in dataframe.}

\item{num_bins}{Integer. Number of bins per histogram plot, default is 30 bins.}
}
\value{
ggplot object
}
\description{
Creates histograms for numerical features within a dataframe using ggplot2
}
