% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EDAhelper.R
\name{preprocess}
\alias{preprocess}
\title{Preprocess data in txt, csv, Excel, etc. by dealing with missing values in numeric columns.}
\usage{
preprocess(
  path,
  method = NULL,
  fill_value = NULL,
  read_func = readr::read_csv,
  ...
)
}
\arguments{
\item{path}{string, path of data file.}

\item{method}{\code{c(NULL, 'most_frequent', 'mean', 'median', 'constant')},  default=\code{NULL}

The imputation method.

If \code{NULL}, then missing values are treated as NaN.

If 'mean', then replace missing values using the mean along each column.

If 'median', then replace missing values using the median along each column.

If 'most_frequent', then replace missing using the most frequent value along each column. If there is more than one such value, only the smallest is returned.

If 'constant', then replace missing values with fill_value.}

\item{fill_value}{\code{NULL} or numeric_value, default=\code{NULL}

When method='constant', fill_value is used to replace all occurrences of missing values.
If left to the default, fill_value will be 0 when imputing numerical data.}

\item{read_func}{\verb{readr::read_*} function name, default=\code{readr::read_csv}}

\item{...}{Any keyword arguments are defined in read_func}
}
\value{
a tibble
}
\description{
Preprocess data in txt, csv, Excel, etc. by dealing with missing values in numeric columns.
}
\examples{
preprocess(readr::readr_example("mtcars.csv"))
}
