% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/column_stats.R
\name{column_stats}
\alias{column_stats}
\title{Obtain summary statistics of column(s) including count, mean, median, mode, Q1, Q3,
variance, standard deviation, correlation, and covariance in table format.}
\usage{
column_stats(data, columns)
}
\arguments{
\item{data}{A data set from which columns will be selected}

\item{columns}{A list of columns for which to obtain summary stats, correlation matrix, and covariance matrix
(if > 1 column arguments used)}
}
\value{
three pd.DataFrames
}
\description{
Obtain summary statistics of column(s) including count, mean, median, mode, Q1, Q3,
variance, standard deviation, correlation, and covariance in table format.
}
\examples{
column_stats(iris, c('Sepal.Length', 'Sepal.Width', 'Petal.Length'))
}
