% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/numeric_plots.R
\name{numeric_plots}
\alias{numeric_plots}
\title{Generate scattered plot matrix of numeric features for EDA}
\usage{
numeric_plots(df)
}
\arguments{
\item{df}{A data.frame}
}
\value{
a GGally plot object with the scattered plot matrix of numeric
features.
}
\description{
Generate scattered plot matrix of numeric features for EDA
}
\examples{
df <- data.frame(names=c("Vera", "Jennifer", "Rowan", "Steven"),
  scores =c(90, 80, 70, 80),
  age=c(12, 23, 20, 100))

numeric_plots(df)
}
